Code style : refactored private methods into a separate module to improve sseService's readability
