--- TODO.md
@@ -1,9 +1,0 @@
-# TODO
-
-## Features
-
-  - Add middlewares + export built-in middleware for polyfill support
-
-## Technical
-
-  - Add d.ts files
